доделал list-page + мелкие фиксы
